+++ packages/uring_api/src/_uring_api/uring_api_submit.c
@@ -51,8 +51,8 @@ static int parse_ull_arg(PyObject *obj, unsigned long long *value_out) {
 }
 
 static int parse_recv_multishot_args(PyObject *const *args, Py_ssize_t nargs, int *fd_out, PyObject **buf_group_out,
-                                   PyObject **user_data_out, unsigned int *flags_out,
-                                   unsigned long long *base_sequence_out) {
+                                     PyObject **user_data_out, unsigned int *flags_out,
+                                     unsigned long long *base_sequence_out) {
     Py_ssize_t positional_optional_count;
 
     if (nargs < 2) {
@@ -91,8 +91,8 @@ static int parse_recv_multishot_args(PyObject *const *args, Py_ssize_t nargs, in
     return 0;
 }
 
-static int parse_send_args(const char *name, PyObject *const *args, Py_ssize_t nargs, Py_ssize_t max_nargs,
-                           int *fd_out, Py_buffer *view_out, PyObject **user_data_out, unsigned int *flags_out,
+static int parse_send_args(const char *name, PyObject *const *args, Py_ssize_t nargs, Py_ssize_t max_nargs, int *fd_out,
+                           Py_buffer *view_out, PyObject **user_data_out, unsigned int *flags_out,
                            unsigned int *zc_flags_out, int parse_zc_flags) {
     if (nargs < 2) {
         PyErr_Format(PyExc_TypeError, "%s() missing required arguments 'fd' and 'data'", name);
@@ -137,8 +137,7 @@ static int parse_accept_listener_args(const char *name, PyObject *const *args, P
         return -1;
     }
     if (nargs > max_args) {
-        PyErr_Format(PyExc_TypeError, "%s() takes at most %zd positional arguments (%zd given)", name, max_args,
-                     nargs);
+        PyErr_Format(PyExc_TypeError, "%s() takes at most %zd positional arguments (%zd given)", name, max_args, nargs);
         return -1;
     }
     if (parse_socket_fd(args[0], fd_out) < 0) {
@@ -196,7 +195,7 @@ PyObject *UringApiRing_submit_recv_impl(UringApiRing *self, int fd, Py_buffer *v
         } else {
             io_uring_prep_recv(sqe, fd, view->buf, (size_t)view->len, 0);
             sqe_set_completion(self, sqe, completion);
-            if (submit_one(self) < 0) {
+            if (submit_one_completion(self, (PyObject *)completion) < 0) {
                 failed = 1;
             }
         }
@@ -244,7 +243,7 @@ PyObject *UringApiRing_submit_recv_buf_impl(UringApiRing *self, int fd, PyObject
             sqe->flags |= IOSQE_BUFFER_SELECT;
             sqe->buf_group = buf_group->group_id;
             sqe_set_completion(self, sqe, completion);
-            if (submit_one(self) < 0) {
+            if (submit_one_completion(self, (PyObject *)completion) < 0) {
                 failed = 1;
             }
         }
@@ -309,7 +308,7 @@ PyObject *UringApiRing_submit_recv_multishot_impl(UringApiRing *self, int fd, Py
             sqe->flags |= IOSQE_BUFFER_SELECT;
             sqe->buf_group = buf_group->group_id;
             sqe_set_completion(self, sqe, completion);
-            if (submit_one(self) < 0) {
+            if (submit_one_completion(self, (PyObject *)completion) < 0) {
                 failed = 1;
             }
         }
@@ -343,7 +342,7 @@ PyObject *UringApiRing_submit_send_impl(UringApiRing *self, int fd, Py_buffer *v
         } else {
             io_uring_prep_send(sqe, fd, view->buf, (size_t)view->len, (int)flags);
             sqe_set_completion(self, sqe, completion);
-            if (submit_one(self) < 0) {
+            if (submit_one_completion(self, (PyObject *)completion) < 0) {
                 failed = 1;
             }
         }
@@ -383,7 +382,7 @@ PyObject *UringApiRing_submit_read_impl(UringApiRing *self, int fd, Py_buffer *v
         } else {
             io_uring_prep_read(sqe, fd, view->buf, (unsigned)view->len, (__u64)offset);
             sqe_set_completion(self, sqe, completion);
-            if (submit_one(self) < 0) {
+            if (submit_one_completion(self, (PyObject *)completion) < 0) {
                 failed = 1;
             }
         }
@@ -423,7 +422,7 @@ PyObject *UringApiRing_submit_write_impl(UringApiRing *self, int fd, Py_buffer *
         } else {
             io_uring_prep_write(sqe, fd, view->buf, (unsigned)view->len, (__u64)offset);
             sqe_set_completion(self, sqe, completion);
-            if (submit_one(self) < 0) {
+            if (submit_one_completion(self, (PyObject *)completion) < 0) {
                 failed = 1;
             }
         }
@@ -464,7 +463,7 @@ PyObject *UringApiRing_submit_openat_impl(UringApiRing *self, int dfd, PyObject 
             } else {
                 io_uring_prep_openat(sqe, dfd, path_state->path, flags, mode);
                 sqe_set_completion(self, sqe, completion);
-                if (submit_one(self) < 0) {
+                if (submit_one_completion(self, (PyObject *)completion) < 0) {
                     failed = 1;
                 }
             }
@@ -511,7 +510,7 @@ PyObject *UringApiRing_submit_statx_impl(UringApiRing *self, int dfd, PyObject *
             } else {
                 io_uring_prep_statx(sqe, dfd, statx_state->path, flags, mask, (struct statx *)statx_state->view.buf);
                 sqe_set_completion(self, sqe, completion);
-                if (submit_one(self) < 0) {
+                if (submit_one_completion(self, (PyObject *)completion) < 0) {
                     failed = 1;
                 }
             }
@@ -546,7 +545,7 @@ PyObject *UringApiRing_submit_send_zc_impl(UringApiRing *self, int fd, Py_buffer
         } else {
             io_uring_prep_send_zc(sqe, fd, view->buf, (size_t)view->len, (int)flags, zc_flags);
             sqe_set_completion(self, sqe, completion);
-            if (submit_one(self) < 0) {
+            if (submit_one_completion(self, (PyObject *)completion) < 0) {
                 failed = 1;
             }
         }
@@ -592,7 +591,7 @@ PyObject *UringApiRing_submit_sendto_impl(UringApiRing *self, int fd, Py_buffer 
             io_uring_prep_sendto(sqe, fd, sendto_state->view.buf, (size_t)sendto_state->view.len, (int)flags,
                                  (struct sockaddr *)&sendto_state->addr, sendto_state->addrlen);
             sqe_set_completion(self, sqe, completion);
-            if (submit_one(self) < 0) {
+            if (submit_one_completion(self, (PyObject *)completion) < 0) {
                 failed = 1;
             }
         }
@@ -633,7 +632,7 @@ PyObject *UringApiRing_submit_recvmsg_impl(UringApiRing *self, int fd, Py_buffer
         } else {
             io_uring_prep_recvmsg(sqe, fd, &msg_state->msg, 0);
             sqe_set_completion(self, sqe, completion);
-            if (submit_one(self) < 0) {
+            if (submit_one_completion(self, (PyObject *)completion) < 0) {
                 failed = 1;
             }
         }
@@ -683,7 +682,7 @@ PyObject *UringApiRing_submit_sendmsg_impl(UringApiRing *self, int fd, Py_buffer
         } else {
             io_uring_prep_sendmsg(sqe, fd, &msg_state->msg, flags);
             sqe_set_completion(self, sqe, completion);
-            if (submit_one(self) < 0) {
+            if (submit_one_completion(self, (PyObject *)completion) < 0) {
                 failed = 1;
             }
         }
@@ -733,7 +732,7 @@ PyObject *UringApiRing_submit_sendmsg_zc_impl(UringApiRing *self, int fd, Py_buf
         } else {
             io_uring_prep_sendmsg_zc(sqe, fd, &msg_state->msg, flags);
             sqe_set_completion(self, sqe, completion);
-            if (submit_one(self) < 0) {
+            if (submit_one_completion(self, (PyObject *)completion) < 0) {
                 failed = 1;
             }
         }
@@ -767,7 +766,7 @@ PyObject *UringApiRing_submit_accept_impl(UringApiRing *self, int fd, unsigned i
         } else {
             io_uring_prep_accept(sqe, fd, NULL, NULL, flags);
             sqe_set_completion(self, sqe, completion);
-            if (submit_one(self) < 0) {
+            if (submit_one_completion(self, (PyObject *)completion) < 0) {
                 failed = 1;
             }
         }
@@ -781,8 +780,8 @@ PyObject *UringApiRing_submit_accept_impl(UringApiRing *self, int fd, unsigned i
     return Py_NewRef(completion);
 }
 
-PyObject *UringApiRing_submit_accept_multishot_impl(UringApiRing *self, int fd, unsigned int flags,
-                                                    PyObject *user_data, unsigned long long base_sequence) {
+PyObject *UringApiRing_submit_accept_multishot_impl(UringApiRing *self, int fd, unsigned int flags, PyObject *user_data,
+                                                    unsigned long long base_sequence) {
     struct io_uring_sqe *sqe;
     PyObject *completion = NULL;
     UringApiCompletion *pending;
@@ -808,7 +807,7 @@ PyObject *UringApiRing_submit_accept_multishot_impl(UringApiRing *self, int fd, 
              * let callers use getpeername() on the accepted fd when needed. */
             io_uring_prep_multishot_accept(sqe, fd, NULL, NULL, flags);
             sqe_set_completion(self, sqe, completion);
-            if (submit_one(self) < 0) {
+            if (submit_one_completion(self, (PyObject *)completion) < 0) {
                 failed = 1;
             }
         }
@@ -853,7 +852,7 @@ PyObject *UringApiRing_submit_connect_impl(UringApiRing *self, int fd, PyObject 
         } else {
             io_uring_prep_connect(sqe, fd, (struct sockaddr *)&sockaddr_state->addr, sockaddr_state->addrlen);
             sqe_set_completion(self, sqe, completion);
-            if (submit_one(self) < 0) {
+            if (submit_one_completion(self, (PyObject *)completion) < 0) {
                 failed = 1;
             }
         }
@@ -887,7 +886,7 @@ PyObject *UringApiRing_submit_poll_impl(UringApiRing *self, int fd, unsigned int
         } else {
             io_uring_prep_poll_add(sqe, fd, poll_mask);
             sqe_set_completion(self, sqe, completion);
-            if (submit_one(self) < 0) {
+            if (submit_one_completion(self, (PyObject *)completion) < 0) {
                 failed = 1;
             }
         }
@@ -923,7 +922,7 @@ PyObject *UringApiRing_submit_poll_multishot_impl(UringApiRing *self, int fd, un
         } else {
             io_uring_prep_poll_multishot(sqe, fd, poll_mask);
             sqe_set_completion(self, sqe, completion);
-            if (submit_one(self) < 0) {
+            if (submit_one_completion(self, (PyObject *)completion) < 0) {
                 failed = 1;
             }
         }
@@ -979,7 +978,7 @@ PyObject *UringApiRing_submit_poll_remove_impl(UringApiRing *self, PyObject *tar
         } else {
             io_uring_prep_poll_remove(sqe, (unsigned long long)(uintptr_t)target_completion);
             sqe_set_completion(self, sqe, (PyObject *)completion);
-            if (submit_one(self) < 0) {
+            if (submit_one_completion(self, (PyObject *)completion) < 0) {
                 failed = 1;
             }
         }
@@ -1017,7 +1016,7 @@ PyObject *UringApiRing_submit_cancel_impl(UringApiRing *self, PyObject *target_c
         } else {
             io_uring_prep_cancel(sqe, target_completion, 0);
             sqe_set_completion(self, sqe, (PyObject *)completion);
-            if (submit_one(self) < 0) {
+            if (submit_one_completion(self, (PyObject *)completion) < 0) {
                 failed = 1;
             }
         }
@@ -1051,7 +1050,7 @@ PyObject *UringApiRing_submit_shutdown_impl(UringApiRing *self, int fd, int how,
         } else {
             io_uring_prep_shutdown(sqe, fd, how);
             sqe_set_completion(self, sqe, completion);
-            if (submit_one(self) < 0) {
+            if (submit_one_completion(self, (PyObject *)completion) < 0) {
                 failed = 1;
             }
         }
@@ -1085,7 +1084,7 @@ PyObject *UringApiRing_submit_close_impl(UringApiRing *self, int fd, PyObject *u
         } else {
             io_uring_prep_close(sqe, fd);
             sqe_set_completion(self, sqe, completion);
-            if (submit_one(self) < 0) {
+            if (submit_one_completion(self, (PyObject *)completion) < 0) {
                 failed = 1;
             }
         }
@@ -1120,7 +1119,7 @@ PyObject *UringApiRing_submit_socket_impl(UringApiRing *self, int domain, int ty
         } else {
             io_uring_prep_socket(sqe, domain, type, protocol, flags);
             sqe_set_completion(self, sqe, completion);
-            if (submit_one(self) < 0) {
+            if (submit_one_completion(self, (PyObject *)completion) < 0) {
                 failed = 1;
             }
         }
@@ -1227,7 +1226,7 @@ PyObject *UringApiRing_submit_statx_fdsize_impl(UringApiRing *self, int fd, PyOb
                 io_uring_prep_statx(sqe, fd, "", URING_API_AT_EMPTY_PATH, URING_API_STATX_SIZE_MASK,
                                     (struct statx *)statx_fdsize_state->buf);
                 sqe_set_completion(self, sqe, completion);
-                if (submit_one(self) < 0) {
+                if (submit_one_completion(self, (PyObject *)completion) < 0) {
                     failed = 1;
                 }
             }